Added networkStatus.js and networkStatus.test.js components wrote some tests
