+++ src/components/identicon.js
@@ -4,7 +4,7 @@ import Jazzicon, { jsNumberForAddress } from 'react-jazzicon'
 class IdentIcon extends Component {
  render() {
     return (
-      <Jazzicon diameter={35} seed={jsNumberForAddress('0x1111111111111111111111111111111111111111')} />
+      <Jazzicon diameter={32} seed={jsNumberForAddress('0x1111111111111111111111111111111111111111')} />
     );
   }
 }
