+++ src/components/navbar.js
@@ -1,19 +1,12 @@
 import React from 'react';
 import IdentIcon from './identicon';
+import NetworkStatus from './networkStatus';
 import { makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
-import AccountCircle from '@material-ui/icons/AccountCircle';
-import Switch from '@material-ui/core/Switch';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import FormGroup from '@material-ui/core/FormGroup';
-import MenuItem from '@material-ui/core/MenuItem';
-import Menu from '@material-ui/core/Menu';
-import Link from '@material-ui/core/Link';
-
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -27,86 +20,24 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function Navbar({account, connect, disconnect}) {
+export default function Navbar({account, connect, disconnect, auth}) {
   const classes = useStyles();
-  const [auth, setAuth] = React.useState(true);
-  const [anchorEl, setAnchorEl] = React.useState(null);
-  const open = Boolean(anchorEl);
-
-  const onConnect = (event) => {
-    connect();
-    setAuth(true);
-  };
-
-  const onDisconnect = (event) => {
-    disconnect();
-    setAuth(false);
-  };
-
-  const handleMenu = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
     return (
     <div className={classes.root}>
       <AppBar position="static" color="inherit">
-        <Toolbar>
-          <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
-            <MenuIcon />
-          </IconButton>
-           <Typography variant="h6" className={classes.title}>
-            ReFi
-          </Typography>
-              <div class="account_address">
-              {account}
-              <IconButton
-                aria-label="account of current user"
-                aria-controls="menu-appbar"
-                aria-haspopup="true"
-                onClick={handleMenu}
-                color="inherit"
-              >
-              <IdentIcon />
-              <Menu
-                id="menu-appbar"
-                anchorEl={anchorEl}
-                anchorOrigin={{
-                  vertical: 'top',
-                  horizontal: 'right',
-                }}
-                keepMounted
-                transformOrigin={{
-                  vertical: 'top',
-                  horizontal: 'right',
-                }}
-                open={open}
-                onClose={handleClose}
-                >
-                {auth?
-                <Link
-                component="button"
-                variant="body2"
-                onClick={onDisconnect}
-                >
-                Disconnect
-                </Link>
-
-                :<Link
-                component="button"
-                variant="body2"
-                onClick={onConnect}
-                >
-                Connect
-                </Link>
-              }
-              </Menu>
-               </IconButton>
-            </div>
-        </Toolbar>
+      <Toolbar>
+      <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
+        <MenuIcon />
+      </IconButton>
+      <Typography variant="h6" className={classes.title}>
+        ReFi
+      </Typography>
+        <IdentIcon />
+        <div className="account_address">
+          {account}
+        <NetworkStatus connect={connect} disconnect={disconnect} auth={auth} />
+      </div>
+      </Toolbar>
       </AppBar>
     </div>
   );
