+++ src/index.js
@@ -1,7 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Web3 from 'web3';
-
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
